requirements.txt
datasets
huggingface_hub
matplotlib
pandas
pillow
scikit-learn
scipy
torch
torchvision
tqdm